Move pages to their category and rename some pages
