--- docs/developer/introduction.md
@@ -1,8 +1,0 @@
-### Introduction
-
-Welcome to the Pumpkin Documentation!
-
-Whether you're an internal Pumpkin developer or working on a Pumpkin plugin, this documentation is your resource for everything Pumpkin.
-
-#### Plugin Development
-For a better documentation structure, plugin development docs have been moved to a [separate category](/plugin-dev/introduction)
