+++ docs/development/contributing/networking/authentication.md
