+++ docs/guide/config/basic.md
