+++ docs/guide/config/compression.md
