+++ docs/guide/config/index.md
@@ -1,3 +1,7 @@
+---
+title: Introduction
+---
+
 ### Configuration
 
 Pumpkin offers a robust configuration system that allows users to customize various aspects of the server's behavior without relying on external plugins. This provides flexibility and control over the server's operation.
