+++ docs/guide/config/resource-pack.md
