--- docs/plugin-dev/plugin-template/introduction.md
@@ -1,9 +1,0 @@
-# Writing an Example Plugin
-This tutorial will guide you through creating a new plugin.
-
-This tutorial expects that you have some experience with rust and with using a command line.
-
-Topics described in this tutorial:
-- Modifying the join message
-- Creating a rock-paper-scissors command (coming soon)
-- Saving plugin data in the plugin's directory (coming soon)
